1.处理各种编译警告，还剩下二十多个警告。2.解决不同版本windows对配置文件的影响，添加删除隐式的换行字符的问题。3.修改界面显示中文字符乱码的问题

+++ CarSeat_Recognization/CarSeat_Recognization/common/utils.h
@@ -48,7 +48,7 @@ namespace utils
 	int parseIp(const char *content, const char *name);
 	
 	// XXX={"AAA":"MMM", "BBB":"NNN", "CCC":"LLL"}
-	int parseMap(const char*content, const char*name, std::unordered_map<std::string, std::string>*pMap);
+	size_t parseMap(const char*content, const char*name, std::unordered_map<std::string, std::string>*pMap);
 
 	// XXX=AAA
 	bool getValueByName(const char *content, const char*name, char*value);
